[TASK] WIP - Separation of audio processing for recording
